Sidebar - Changed layout and styles
GoogleCloud - hide by default.
Fixed the "Drag components to the canvas" message and made it more prominent.
Moved the Input and Output nodes to the Special section collapsed by default.
Moded the exporters to the Debug section collapsed by default.
Removed the TFX section in the component library.
Unified the styles a bit.