vertexAiCompiler - Stopped setting the pipeline context name to the component name
After the recent changes, some pipelines/components started having custom names and that started affecting the cache reuse.
I want the cache to be global, so I've made this change.
It's still possible to pass the context name explicitly.

I also set the pipelineJob.displayName to the component name, but it does not seem to have any effect.