feat: Added user data migration to new website
Migrating the local data (e.g. IndexedDB) between domains is really hard due to CORS. Only page loaded form some origin can access the data for that origin.
GitHub custom domain support make it even harder. When developer adds custom domain to the GitHub pages repository, GitHub sets up hard permanent redirect from the original user.github.io domain. This makes the data stored for that domain completely inaccessible.
I had to create a separate repo serving the new domain just to prevent GitHub redirects on the old domain.

Migration design.
The only way to communicate between cross-origin pages is to use postMessage.
When the user visits the old website normally, they're redirected to the new website.
However I made it possible to put the website in the migration source mode where it sends out the user data that should be migrated. In this case the redirect does not occur.
When the user visits the new website, the code performs the migration if it hasn't been done before. The code dynamically opens an IFrame with the old website in the migration source mode and listens for incoming message with the migrated data. Upon receiving the message the code writes the files to the DB and adds a migration completion marker. This way the migration only happens once (successfully).