refactor: VertexAiCompiler - Refactored the compiler to change how the executors, components and tasks are added during the compilation
Preparing for multi-DAG support.
Previously the `taskSpecToVertexTaskSpecComponentSpecAndExecutorSpec` function would return a set of execotor, component and one or more tasks and the caller had to link them together. This refactoring changes that so that the executors and components are registered immediately. This is needed for multi-DAG compilation where task IDs of multiple DAGs are not longer unique and so it's not possible to derive unique executor and component IDs from the task IDs.

Simplified the parameter-to-artifact converter task generation.
I had to add de-duplication for such tasks, which already results in less tasks in the compiled pipeline. Full de-duplication fo executors, components and tasks will come next.