fix: VertexAiCompiler - Fixed the way the default arguments are compiled
The bug was introduced in a recent refactoring commit: "refactor: VertexAiCompiler - Changed the way resolveCommandLine detects parameter arguments".

The bug caused default arguments to be considered artifacts which led to the compiler inserting many "Make artifact" tasks to convert the constant default values to the artifacts that the components incorrectly expected.

It's interesting that the compiled pipelines still worked (albeit with extra unnecessary conversion steps).