refactor: Functions now perform downloading using a passed downloader function
This refactoring improves the code sharing with the VSCode extension by making the downloading and caching routines configurable from the top of the app.
This dependency injection is also making the functions easier to test.

Additional feature: The caching strategy (whether to try downloading from Web before using the cache) is now dependent on the URLL: The data for the immutable hash-based GitHub and GitLab URLs is now always taken from cahce first while everything else is downloaded from the Web first.