refactor: Switched from downloadText functions to downloadData functions that accept transformers
The introduction of transformers makes it possible to validate the data before caching.
This change makes it easier to download and cache binary data.
Transformers are not just used for the purpose of transforming the data. They are also used to validate the downloaded data before putting it into cache.
(It's still possible to first put invalid data into cache with a permissable transformer and then get errors when getting that data from cache later. But that was also possible before.)